// libs/jsonresume-theme-solid/gen/schema.d.ts
export type Iso8601 = string;

export interface ResumeSchema {
// ...
  education?: {
    /**
     * e.g. Massachusetts Institute of Technology
     */
    institution?: string;
    /**
     * e.g. http://facebook.example.com
     */
    url?: string;
    /**
     * e.g. Arts
     */
    area?: string;
    /**
     * e.g. Bachelor
     */
    studyType?: string;
    startDate?: Iso8601;
    endDate?: Iso8601;
    /**
     * grade point average, e.g. 3.67/4.0
     */
    score?: string;
    /**
     * List notable courses/subjects
     */
    courses?: string[];
    [k: string]: unknown;
  }[];
// ...
}